refactoring: split cpp module into class, enum and interface
